Former-commit-id: a8070f5cc7a56efb4857d8490be2af55eba9d4f8

+++ web/concrete/models/block.php
@@ -225,8 +225,10 @@ class Block extends Object {
 		$scrapbookHelper=Loader::helper('concrete/scrapbook'); 
 		$globalScrapbookC = $scrapbookHelper->getGlobalScrapbookPage(); 
 		
-		if( $this->cID==$globalScrapbookC->cID ) return true;
-		
+		if( $this->cID==$globalScrapbookC->cID ) {
+			$this->bIsGlobal = true;	
+			return true;
+		}
 		$q = "SELECT b.bID FROM Blocks AS b, CollectionVersionBlocks AS cvb ".
 			 "WHERE b.bID = '{$this->bID}' AND cvb.bID=b.bID AND cvb.cID=".intval($globalScrapbookC->getCollectionId())." LIMIT 1";
 			 
@@ -319,10 +321,15 @@ class Block extends Object {
 		
 		// ONLY ALLOWS ITEMS THAT START WITH "action_";
 		
-		return @call_user_func_array(array($bc, $method), array());
+		return @$bc->{$method}();
 	}
 	
-	public function getInstance() {
+	public function getInstance() {		
+		if ($this->instance->cacheBlockRecord() && is_object($this->instance->getBlockControllerData())) {
+			$this->instance->__construct();
+		} else {
+			$this->instance = Loader::controller($this);
+		}		
 		return $this->instance;
 	}
 	
@@ -807,6 +814,9 @@ class Block extends Object {
 			$q = "delete from CollectionVersionBlocks where bID = '$bID'";
 			$r = $db->query($q);
 
+			$q = "delete from ComposerContentLayout where bID = '$bID'";
+			$r = $db->query($q);
+
 			$q = "delete from CollectionVersionBlockPermissions where bID = '$bID'";
 			$r = $db->query($q);
 			
@@ -1093,21 +1103,29 @@ class Block extends Object {
 	 * Removes a cached version of the block 
 	 */
 	public function refreshCache() {
-		$c = $this->getBlockCollectionObject();
-		$a = $this->getBlockAreaObject();
-		if (is_object($c) && is_object($a)) {
-			Cache::delete('block', $this->getBlockID() . ':' . $c->getCollectionID() . ':' . $c->getVersionID() . ':' . $a->getAreaHandle());
-			Cache::delete('block_view_output', $c->getCollectionID() . ':' . $this->getBlockID() . ':' . $a->getAreaHandle());
-			Cache::delete('collection_blocks', $c->getCollectionID() . ':' . $c->getVersionID());
+		// if the block is a global block, we need to delete all cached versions that reference it.
+		if ($this->bIsGlobal) {
+			$this->refreshCacheAll();
+		} else { 
+			$c = $this->getBlockCollectionObject();
+			$a = $this->getBlockAreaObject();
+			if (is_object($c) && is_object($a)) {
+				Cache::delete('block', $this->getBlockID() . ':' . $c->getCollectionID() . ':' . $c->getVersionID() . ':' . $a->getAreaHandle());
+				Cache::delete('block_view_output', $c->getCollectionID() . ':' . $this->getBlockID() . ':' . $a->getAreaHandle());
+				Cache::delete('collection_blocks', $c->getCollectionID() . ':' . $c->getVersionID());
+			}
+			Cache::delete('block', $this->getBlockID());		
 		}
-		Cache::delete('block', $this->getBlockID());		
 	}
 	
-	public function refreshCacheAll(){
+	public function refreshCacheAll() {
 		$db = Loader::db();
 		$rows=$db->getAll( 'SELECT cID, cvID, arHandle FROM CollectionVersionBlocks WHERE bID='.intval($this->getBlockID()) ); 
 		foreach($rows as $row){
 			Cache::delete('block', $this->getBlockID() . ':' . intval($row['cID']) . ':' . intval($row['cvID']) . ':' . $row['arHandle'] );
+			Cache::delete('block_view_output', $row['cID'] . ':' . $this->getBlockID(). ':' . $row['arHandle']);
+			Cache::delete('collection_blocks', $row['cID'] . ':' . $row['cvID']);
+			Cache::delete('block', $this->getBlockID());
 		}
 	}
